# openapi.yaml
openapi: 3.1.1
info:
  title: Ayllu.Web | v1
  version: 1.0.0
servers:
  - url: https://ayllu.local/
paths:
...
  /api/v1/dialectics/{dialecticId}/thesis:
    post:
      tags:
        - Dialectics
      summary: Publica uma tese associada à dialética especificada.
      parameters:
        - name: dialecticId
          in: path
          description: O identificador único da dialética cuja tese será publicada. Não
            pode ser nulo ou vazio.
          required: true
          schema:
            type: string
      requestBody:
        description: Token de cancelamento da operação.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThesisRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateThesisRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateThesisRequest"
        required: true
      responses:
        "200":
          description: Uma response indicando o resultado da operação com HTTP status code
            200 OK.
  /api/v1/dialectics/{dialecticId}/thesis/{thesisId}:
    put:
      tags:
        - Dialectics
      summary: Atualiza uma tese associada à uma dialética específica.
      parameters:
        - name: dialecticId
          in: path
          description: O identificador único da dialética cuja tese será atualizada. Não
            pode ser nulo ou vazio.
          required: true
          schema:
            type: string
        - name: thesisId
          in: path
          description: O identificador único da tese a ser atualizada.
          required: true
          schema:
            type: string
      requestBody:
        description: Token de cancelamento da operação.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThesisRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateThesisRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateThesisRequest"
        required: true
      responses:
        "200":
          description: Uma response HTTP 200 Ok com o resultado da atualização.
components:
  schemas:
...
    CreateThesisRequest:
      required:
        - content
      type: object
      properties:
        content:
          type: string
...
    UpdateThesisRequest:
      required:
        - content
      type: object
      properties:
        content:
          type: string
...